# backend/swagger.yaml
swagger: "2.0"
info:
  title: KnightWise
  description: |
    API documentation for KnightWise, the computer science guided learning platform.

    ## Question Types and Grading
    
    KnightWise supports five non-programming question types, each with its own grading algorithm:
    
    ### Multiple Choice
    - **Input Format**: String (e.g., `"Paris"`)
    - **Grading**: Exact match (case-insensitive, trimmed)
    - **Normalized Score**: 1.0 (correct) or 0.0 (incorrect)
    
    ### Fill-In-the-Blanks
    - **Input Format**: String (e.g., `"Paris"`)
    - **Grading**: Exact or fuzzy matching using Levenshtein distance
    - **Normalized Score**: 0.0 to 1.0 based on similarity (1 - distance/maxLength)
    - **Partial Credit**: Yes (>= 50% similarity, API feedback says "Almost correct!")
    
    ### Select-All-That-Apply
    - **Input Format**: Array of strings (e.g., `["Python", "JavaScript"]`)
    - **Grading**: (correct selections - incorrect selections) / total correct options
    - **Normalized Score**: 0.0 to 1.0 based on how many correct selections
    - **Partial Credit**: Yes
    
    ### Ranked Choice
    - **Input Format**: Array of strings in ranked order (e.g., `["First", "Second", "Third"]`)
    - **Grading**: Kendall tau distance (counts pairwise inversions)
    - **Normalized Score**: 0.0 to 1.0, based on 1.0 - (inversions / maxInversions)
    - **Partial Credit**: Yes
    
    ### Drag-and-Drop
...
    - **Grading**: Correct placements / total items
...
    
...
    
...
paths:
...
  /users/search:
    get:
      tags:
        - Users
      summary: Search users by username.
      operationId: searchUsers
      description: |
        Returns a paginated list of users whose username partially matches the query string.
        - Match is case-insensitive. Empty or omitted username returns all users.
        - Invalid or out-of-range page values are clamped to a valid page rather than rejected.
      security:
        - BearerAuth: []
      parameters:
        - name: username
          in: query
          required: false
          type: string
          description: Partial, case-insensitive username to search for. Omit or leave empty to return all users.
          example: knight
        - name: page
          in: query
          required: false
          type: integer
          description: Page number (default 1, PAGE_SIZES.USER_SEARCH per page, clamped to valid range).
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/UserSearchResponse'
        401:
          description: Unauthorized
        500:
          description: Server Error
...
  /users/{id}:
    get:
      tags:
        - Users
      summary: Fetch user profile info.
      operationId: getUserInfo
      description: Retrieves a user's profile information including currency, experience, and currently equipped store items.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          type: integer
          description: The unique ID of the user.
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/UserInfoResponse'
        400:
          description: Bad Request - Invalid user ID
        401:
          description: Unauthorized
        404:
          description: User Not Found
        500:
          description: Server Error
    put:
      tags:
        - Users
      summary: Update user profile.
      operationId: updateUser
      description: Updates editable profile fields (first name, last name). Can only be performed by the account owner.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          type: integer
          description: The unique ID of the user.
        - in: body
          name: updateUserItem
          description: Updated profile fields.
          schema:
            $ref: '#/definitions/UpdateUser'
      responses:
        200:
          description: OK
        400:
          description: Bad Request - Invalid user ID or name validation failed
        401:
          description: Unauthorized
        403:
          description: Forbidden - Can only update own profile
        404:
          description: User Not Found
        500:
          description: Server Error
    delete:
      tags:
        - Users
      summary: Delete user account.
      operationId: deleteAccount
      description: Deletes a user's account and all associated data. Can only be performed by the account owner.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          type: integer
          description: The unique ID of the user.
      responses:
        200:
          description: OK
        400:
          description: Bad Request - Invalid user ID
        401:
          description: Unauthorized
        403:
          description: Forbidden - Can only delete own account
        404:
          description: User Not Found
        500:
          description: Server Error
  /users/{id}/purchases:
    get:
      tags:
        - Users
      summary: Fetch user's purchased store items.
      operationId: getPurchases
      description: Retrieves all store items purchased by the user, including equip state. Any authenticated user can view any user's purchases.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          type: integer
          description: The unique ID of the user.
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/PurchasesResponse'
        400:
          description: Bad Request - Invalid user ID
        401:
          description: Unauthorized
        500:
          description: Server Error
  /users/{id}/equip:
    put:
      tags:
        - Users
      summary: Equip a purchased store item.
      operationId: equipItem
      description: |
        Equips a purchased store item for the user. Enforces equip limits per item type:
        - **Flair**: Max 3 equipped at once. Returns 409 if limit reached.
        - **Profile Picture**: Max 1 equipped at once. Automatically unequips the current profile picture when a new one is equipped.

        Can only be performed by the account owner.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          type: integer
          description: The unique ID of the user.
        - in: body
          name: equipItem
          description: Item to equip.
          schema:
            $ref: '#/definitions/EquipItem'
      responses:
        200:
          description: OK
        400:
          description: Bad Request - Invalid user ID or item already equipped
        401:
          description: Unauthorized
        403:
          description: Forbidden - Can only equip for own account
        404:
          description: Item not purchased
        409:
          description: Conflict - Equip limit reached for item type
        500:
          description: Server Error
  /users/{id}/unequip:
    put:
      tags:
        - Users
      summary: Unequip a purchased store item.
      operationId: unequipItem
      description: Unequips a currently equipped store item for the user. Can only be performed by the account owner.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          type: integer
          description: The unique ID of the user.
        - in: body
          name: unequipItem
          description: Item to unequip.
          schema:
            $ref: '#/definitions/EquipItem'
      responses:
        200:
          description: OK
        400:
          description: Bad Request - Invalid user ID or item not equipped
        401:
          description: Unauthorized
        403:
          description: Forbidden - Can only unequip for own account
        404:
          description: Item not purchased
        500:
          description: Server Error
  /users/{id}/follow:
    post:
      tags:
        - Users
      summary: Follow a user.
      operationId: followUser
      description: Follows the user with the given ID. Following is immediate and one-way, no acceptance required from the followee. A user cannot follow themselves or follow the same user twice.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          type: integer
          description: The unique ID of the user to follow.
      responses:
        200:
          description: OK
        400:
          description: Bad Request - Invalid user ID, user already followed, or attempting to follow yourself
        401:
          description: Unauthorized
        404:
          description: User Not Found
        500:
          description: Server Error
    delete:
      tags:
        - Users
      summary: Unfollow a user.
      operationId: unfollowUser
      description: Unfollows the user with the given ID. One-way, does not affect whether the followee follows the requesting user.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          type: integer
          description: The unique ID of the user to unfollow.
      responses:
        200:
          description: OK
        400:
          description: Bad Request - Invalid user ID or user not currently followed
        401:
          description: Unauthorized
        404:
          description: User Not Found
        500:
          description: Server Error
  /users/{id}/stats-opt-in:
    put:
      tags:
        - Users
      summary: Update stats sharing opt-in.
      operationId: updateStatsOptIn
      description: |
        Sets whether the user's response data is included in professor-facing aggregate analytics.
        - Explicitly sets the value rather than toggling, so repeated calls with the same value don't change it.
        - Can only be performed by the account owner.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          type: integer
          description: The unique ID of the user.
        - in: body
          name: statsOptInItem
          description: Desired opt-in state.
          schema:
            $ref: '#/definitions/StatsOptIn'
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/StatsOptInResponse'
        400:
          description: Bad Request - Invalid user ID or optIn is not a boolean
        401:
          description: Unauthorized
        403:
          description: Forbidden - Can only update own settings
        404:
          description: User Not Found
        500:
          description: Server Error
...
securityDefinitions:
  BearerAuth:
    type: apiKey
    name: Authorization
    in: header
    description: "Include the token in the Authorization header as 'Bearer your-token'"
definitions:
...
  UpdateUser:
    type: object
    required:
      - newFirstName
      - newLastName
    properties:
      newFirstName:
        type: string
        example: "Jean-Jacques"
        description: "Allows letters, spaces, hyphens, apostrophes, periods, and accented/Unicode characters."
      newLastName:
        type: string
        example: "O'Malley"
        description: "Allows letters, spaces, hyphens, apostrophes, periods, and accented/Unicode characters."
  UserInfoResponse:
    type: object
    properties:
      user:
        type: object
        properties:
          ID:
            type: integer
            example: 1
          USERNAME:
            type: string
            example: "RickL"
            nullable: true
          FIRSTNAME:
            type: string
            example: "Rick"
            nullable: true
          LASTNAME:
            type: string
            example: "Leinecker"
            nullable: true
          LIFETIME_EXP:
            type: number
            example: 1250.0
          WEEKLY_EXP:
            type: number
            example: 300.0
          DAILY_EXP:
            type: number
            example: 57.5
          COINS:
            type: number
            example: 450.0
      equippedItems:
        type: array
        items:
          $ref: '#/definitions/StoreItem'
  UserSearchResponse:
    type: object
    properties:
      users:
        type: array
        items:
          type: object
          properties:
            ID:
              type: integer
              example: 42
            USERNAME:
              type: string
              example: "knightking"
            FIRSTNAME:
              type: string
              example: "Arthur"
              nullable: true
            LASTNAME:
              type: string
              example: "Pendragon"
              nullable: true
      pagination:
        type: object
        properties:
          page:
            type: integer
            example: 1
            description: The actual page returned (clamped to valid range).
          pageSize:
            type: integer
            example: 20
            description: Number of results per page.
          totalUsers:
            type: integer
            example: 3
            description: Total number of matching users across all pages.
          totalPages:
            type: integer
            example: 1
            description: Total number of pages for this query.
  PurchasesResponse:
    type: object
    properties:
      purchases:
        type: array
        items:
          $ref: '#/definitions/PurchasedItem'
...
  StoreItem:
    type: object
    properties:
      ID:
        type: integer
        example: 1
      TYPE:
        type: string
        example: "flair"
        description: "Item type. One of: flair, profile_picture, background"
      COST:
        type: string
        example: "50.00"
        description: "Returned as string due to MySQL DECIMAL serialization."
      NAME:
        type: string
        example: "Greatest of all Time"
      IS_GUILD_ITEM:
        type: integer
        example: 0
        description: "1 if this is a guild store item, 0 if it is a user store item."
...
  EquipItem:
    type: object
    required:
      - itemId
    properties:
      itemId:
        type: integer
        example: 1
        description: The unique ID of the store item to equip or unequip.
...
  StatsOptIn:
    type: object
    required:
      - optIn
    properties:
      optIn:
        type: boolean
        example: true
        description: True to include this user's responses in aggregate stats, false to exclude.
  StatsOptInResponse:
    type: object
    properties:
      message:
        type: string
        example: "Stats sharing enabled"
      optIn:
        type: boolean
        example: true
...
host: www.knightwise.dev
basePath: /api
schemes:
...